+++ R/formatting_functions.R
@@ -274,9 +274,12 @@ format_sigfig <- function(sigfig, format = "xx", num_fmt = "fg") {
   checkmate::assert_choice(format, c("xx", "xx / xx", "(xx, xx)", "xx - xx", "xx (xx)"))
   function(x, ...) {
     if (!is.numeric(x)) stop("`format_sigfig` cannot be used for non-numeric values. Please choose another format.")
-    num <- formatC(signif(x, digits = sigfig), digits = sigfig, format = num_fmt, flag = "#")
-    num <- gsub("\\.$", "", num) # remove trailing "."
-
+    num <- vapply(x, function(val) {
+      s <- signif(val, digits = sigfig)
+      flag <- if (!is.na(val) && abs(val) >= 0.1) "#" else ""
+      out <- formatC(s, digits = sigfig, format = num_fmt, flag = flag)
+      gsub("\\.$", "", out)
+    }, character(1))
     format_value(num, format)
   }
 }